detach privatesend-client from server and move to wallet
