+++ src/Makefile.am
@@ -100,7 +100,6 @@ BITCOIN_CORE_H = \
   core_io.h \
   core_memusage.h \
   cuckoocache.h \
-  dsnotificationinterface.h \
   fs.h \
   governance.h \
   governance-classes.h \
@@ -126,6 +125,7 @@ BITCOIN_CORE_H = \
   memusage.h \
   merkleblock.h \
   miner.h \
+  module-interface.h \
   net.h \
   net_processing.h \
   netaddress.h \
@@ -140,9 +140,7 @@ BITCOIN_CORE_H = \
   pow.h \
   protocol.h \
   privatesend.h \
-  privatesend-client.h \
   privatesend-server.h \
-  privatesend-util.h \
   random.h \
   reverselock.h \
   reverse_iterator.h \
@@ -189,9 +187,11 @@ BITCOIN_CORE_H = \
   wallet/fees.h \
   wallet/init.h \
   wallet/keepass.h \
+  wallet/privatesend-client.h \
   wallet/rpcwallet.h \
   wallet/wallet.h \
   wallet/walletdb.h \
+  wallet/walletinterface.h \
   wallet/walletutil.h \
   wallet/coinselection.h \
   warnings.h \
@@ -219,7 +219,6 @@ libbitcoin_server_a_SOURCES = \
   chain.cpp \
   checkpoints.cpp \
   consensus/tx_verify.cpp \
-  dsnotificationinterface.cpp \
   httprpc.cpp \
   httpserver.cpp \
   init.cpp \
@@ -238,6 +237,7 @@ libbitcoin_server_a_SOURCES = \
   merkleblock.cpp \
   messagesigner.cpp \
   miner.cpp \
+  module-interface.cpp \
   net.cpp \
   netfulfilledman.cpp \
   net_processing.cpp \
@@ -284,18 +284,18 @@ endif
 libbitcoin_wallet_a_CPPFLAGS = $(AM_CPPFLAGS) $(BITCOIN_INCLUDES)
 libbitcoin_wallet_a_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)
 libbitcoin_wallet_a_SOURCES = \
-  privatesend-client.cpp \
-  privatesend-util.cpp \
   wallet/crypter.cpp \
   wallet/db.cpp \
   wallet/feebumper.cpp \
   wallet/fees.cpp \
   wallet/init.cpp \
   wallet/keepass.cpp \
+  wallet/privatesend-client.cpp \
   wallet/rpcdump.cpp \
   wallet/rpcwallet.cpp \
   wallet/wallet.cpp \
   wallet/walletdb.cpp \
+  wallet/walletinterface.cpp \
   wallet/walletutil.cpp \
   wallet/coinselection.cpp \
   $(BITCOIN_CORE_H)
