+++ src/chaincoind.cpp
@@ -17,11 +17,13 @@
 #include <noui.h>
 #include <util.h>
 #include <masternodeconfig.h>
+#include <module-interface.h>
 #include <httpserver.h>
 #include <httprpc.h>
 #include <utilstrencodings.h>
 #if ENABLE_WALLET
 #include <wallet/init.h>
+#include <wallet/walletinterface.h>
 #endif
 #include <walletinitinterface.h>
 
@@ -67,8 +69,10 @@ bool AppInit(int argc, char* argv[])
 
 #if ENABLE_WALLET
     g_wallet_init_interface.reset(new WalletInit);
+    g_module_interface.reset(new WalletInterface);
 #else
     g_wallet_init_interface.reset(new DummyWalletInit);
+    g_module_interface.reset(new DummyModuleInterface);
 #endif
 
     //
