--- src/dsnotificationinterface.h
@@ -1,30 +1,0 @@
-// Copyright (c) 2015-2018 The Bitcoin Core developers
-// Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
-
-#ifndef BITCOIN_DSNOTIFICATIONINTERFACE_H
-#define BITCOIN_DSNOTIFICATIONINTERFACE_H
-
-#include <validationinterface.h>
-
-class CDSNotificationInterface : public CValidationInterface
-{
-public:
-    CDSNotificationInterface(CConnman* connmanIn): connman(connmanIn) {}
-    virtual ~CDSNotificationInterface() = default;
-
-    // a small helper to initialize current block height in sub-modules on startup
-    void InitializeCurrentBlockTip();
-
-protected:
-    // CValidationInterface
-//    void BlockChecked(const CBlock& block, const CValidationState& state) override;
-//    void NewPoWValidBlock(const CBlockIndex *pindex, const std::shared_ptr<const CBlock> &block) override;
-    void UpdatedBlockTip(const CBlockIndex *pindexNew, const CBlockIndex *pindexFork, bool fInitialDownload) override;
-    void TransactionAddedToMempool(const CTransactionRef& tx) override;
-
-private:
-    CConnman* connman;
-};
-
-#endif // BITCOIN_DSNOTIFICATIONINTERFACE_H
