+++ src/qt/optionsmodel.cpp
@@ -22,12 +22,10 @@
 #ifdef ENABLE_WALLET
 #include <wallet/wallet.h>
 #include <wallet/walletdb.h>
-#endif
+#include <wallet/privatesend-client.h>
 
-#ifdef ENABLE_WALLET
 #include <masternodeconfig.h>
 #endif
-#include <privatesend-client.h>
 
 #include <QNetworkProxy>
 #include <QSettings>
